+++ src/Sepes.Infrastructure/Service/Interface/ISandboxPhaseService.cs
@@ -1,0 +1,10 @@
+﻿using System.Threading;
+using System.Threading.Tasks;
+
+namespace Sepes.Infrastructure.Service.Interface
+{
+    public interface ISandboxPhaseService
+    {
+        Task MoveToNextPhaseAsync(int sandboxId, CancellationToken cancellation = default);
+    }
+}
