+++ src/Sepes.Infrastructure/Service/Interface/ISandboxResourceDeleteService.cs
@@ -1,0 +1,11 @@
+﻿using Sepes.Infrastructure.Dto.Sandbox;
+using System.Threading.Tasks;
+
+namespace Sepes.Infrastructure.Service.Interface
+{
+    public interface ISandboxResourceDeleteService
+    { 
+        //DESTRUCTION
+        Task<SandboxResourceOperationDto> MarkAsDeletedAsync(int resourceId);
+    }
+}
