+++ src/Sepes.Infrastructure/Service/Interface/ISandboxResourceUpdateService.cs
@@ -1,0 +1,22 @@
+﻿using Sepes.Infrastructure.Dto;
+using System.Threading.Tasks;
+
+namespace Sepes.Infrastructure.Service.Interface
+{
+    public interface ISandboxResourceUpdateService
+    {
+        //GENERAL METHODS        
+
+        Task<SandboxResourceDto> Update(int resourceId, SandboxResourceDto updated);   
+
+
+        //MORE SPECIFIC RESOURCE OPERATIONS     
+
+        Task<SandboxResourceDto> UpdateResourceGroup(int resourceId, SandboxResourceDto updated);
+
+
+        Task<SandboxResourceDto> UpdateResourceIdAndName(int resourceId, string azureId, string azureName);
+
+        Task UpdateProvisioningState(int resourceId, string newProvisioningState);
+    }
+}
