+++ src/Sepes.Infrastructure/Service/Interface/ISandboxService.cs
@@ -1,6 +1,7 @@
 ﻿using Sepes.Infrastructure.Constants;
 using Sepes.Infrastructure.Dto.Sandbox;
 using System.Collections.Generic;
+using System.Threading;
 using System.Threading.Tasks;
 
 namespace Sepes.Infrastructure.Service.Interface
@@ -13,10 +14,6 @@ namespace Sepes.Infrastructure.Service.Interface
 
         Task<IEnumerable<SandboxDto>> GetAllForStudy(int studyId);
         Task<SandboxDetailsDto> CreateAsync(int studyId, SandboxCreateDto newSandbox);
-        Task DeleteAsync(int sandboxId);
-
-      
-
-        Task MoveToNextPhase(int sandboxId);
+        Task DeleteAsync(int sandboxId);        
     }
 }
