+++ src/Sepes.Tests/Services/DomainServices/ServiceTestBase.cs
@@ -1,8 +1,8 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
-
 using Sepes.Infrastructure.Model.Context;
 using Sepes.Tests.Constants;
 using Sepes.Tests.Setup;
+using System.Threading.Tasks;
 
 namespace Sepes.Tests.Services
 {
@@ -22,20 +22,20 @@ namespace Sepes.Tests.Services
           return _serviceProvider.GetService<SepesDbContext>();     
         }
 
-        protected SepesDbContext ClearTestDatabase()
+        protected async Task<SepesDbContext> ClearTestDatabase()
         {
             var db = GetDatabase();
-            db.Database.EnsureDeleted();
-            db.Database.EnsureCreated();
+            await db.Database.EnsureDeletedAsync();
+            await db.Database.EnsureCreatedAsync();
 
             return db;
         }
 
-        protected SepesDbContext ClearTestDatabaseAddUser()
+        protected async Task<SepesDbContext> ClearTestDatabaseAddUser()
         {
-            var db = ClearTestDatabase();
+            var db = await ClearTestDatabase();
             db.Users.Add(new Infrastructure.Model.User() { Id = UserConstants.COMMON_CUR_USER_DB_ID, ObjectId = UserConstants.COMMON_CUR_USER_OBJECTID, FullName = UserConstants.COMMON_CUR_USER_FULL_NAME, EmailAddress = UserConstants.COMMON_CUR_USER_EMAIL, UserName = UserConstants.COMMON_CUR_USER_UPN  });
-            db.SaveChangesAsync();
+            await db.SaveChangesAsync();
             return db;
         }
     }
