+++ src/Sepes.Tests/Services/DomainServices/StudySpecificDatasetServiceTests.cs
@@ -18,7 +18,7 @@ namespace Sepes.Tests.Services
         [Fact]
         public async void CreateStudySpecificDataset_WhenStudyIsMissing_ShouldFail()
         {
-            var db = ClearTestDatabase();
+            var db = await ClearTestDatabase();
             var service = DatasetServiceMockFactory.GetStudySpecificDatasetService(_serviceProvider);
             await Assert.ThrowsAsync<NotFoundException>(() => service.CreateStudySpecificDatasetAsync(1, new DatasetCreateUpdateInputBaseDto() { Name = "testds", Location = "norwayeast", Classification = "open" }, "192.168.1.1"));
 
@@ -28,7 +28,7 @@ namespace Sepes.Tests.Services
         [Fact]
         public async void CreateStudySpecificDataset_WhenStudyIsMissingWbs_ShouldFail()
         {
-            var db = ClearTestDatabase();
+            var db = await ClearTestDatabase();
             
             StudyPopulator.Add(db, "Test Study 1", "Vendor for TS1", null, 1);
             await db.SaveChangesAsync();
